fix(core): resolve MSVC-specific _atoi64, pointer return types, and 64-bit MD5 hashing
- Replace MSVC-specific _atoi64 with standard C99/C++11 atoll in ioplayerxml.cpp and monsters.cpp for high-level experience parsing.
- Add fallback compatibility macro #define _atoi64(a) atoll(a) in definitions.h.
- Fix invalid return type in IOMapOTBM::unserializaItemNode (iomapotbm.cpp) by replacing 'return false;' with 'return NULL;' to adhere to C++11 pointer conversion rules.
- Fix 64-bit MD5 calculation by redefining UINT4 as uint32_t instead of unsigned long in md5.h to ensure exact 32-bit word sizes on LP64 architectures (x86_64 / aarch64).
- Include <cmath> in player.cpp to guarantee availability of std::pow across all GCC versions.

+++ source/md5.h
@@ -54,7 +54,8 @@
 
 /* Typedef a 32 bit type */
 #ifndef UINT4
-typedef unsigned long int UINT4;
+#include <stdint.h>
+typedef uint32_t UINT4;
 #endif
 
 /* Data structure for MD5 (Message Digest) computation */
